Create Alice's library.java
Alice is rearranging her library. She takes the innermost shelf and reverses the order of books. She breaks the walls of the shelf. In the end, there will be only books and no shelf walls. Print the order of books.

Input format

The first line contains string 
 displaying her library.

Output format

Print only one string displaying Alice's library after rearrangement.

Note
The first character of the string is '/' and the last character of the string is '\' indicating the outermost walls of the shelf. 

Sample Input
/u/love\i\
Sample Output
iloveu