Create The Squid Game.java
Input:
3
3
3 1 2
5
1 1 1 1 1
6
3 6 4 2 5 1

output:
5
4
20

Explanation:
Test Case 1:

If we choose the first player to be the winner, he will win the game when the second and third players die. Hence, the amount of money won by him will be 
1
+
2
=
3
1+2=3.
If we choose the second player to be the winner, he will win the game when the first and third players die. Hence, the amount of money won by him will be 
3
+
2
=
5
3+2=5.
If we choose the third player to be the winner, he will win the game when the first and second players die. Hence, the amount of money won by him will be 
3
+
1
=
4
3+1=4.
Therefore, we can clearly see that the maximum amount of money that can be won by any player is 
5
5.

Test Case 2: Irrespective of who is chosen, the winner will always win an amount of 
4
4.

Test Case 3: If we choose the sixth player to be the winner, the amount won by him will be 
3
+
6
+
4
+
2
+
5
=
20
3+6+4+2+5=20. It can be proven that if we choose any other player to be the winner, the amount is less than 
20
20. Hence, the maximum amount of money that can be won by any player is 
20
20.